[84] | Feature: Change Price property type to decimal and update related logic in OfficeAttendance and Money classes

+++ src/Domain/Entities/ValueObjects/Money.cs
@@ -12,9 +12,9 @@ public class Money
         ValueAsCents = valueAsCents;
     }
 
-    public static Money Create(long value)
+    public static Money Create(decimal value)
     {
-        var cents = value * 100;
+        var cents = (long)(value * 100);
         return new Money(cents);
     }
